feat(PICO): Make the runner more configurable
Instead of hardcoding the address of `VirtualFree()`, the runner now accepts a pair of ROR13 hashes via the `$TMH` and `$TFH` linker parameters. These are used to resolve a function address which is then hooked with the hardware breakpoint.

When you run `make build`, the "default" runner still hooks `VirtualFree()` as it did before.